+++ src/Channel/resources/database/migrations/2023_07_10_145614_add_currency_to_the_channels_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration {
+    public function up(): void
+    {
+        Schema::table('channels', function (Blueprint $table) {
+            $table->char('currency', 3)->nullable()->after('slug');
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::table('channels', function (Blueprint $table) {
+            $table->dropColumn('currency');
+        });
+    }
+};